refactor: centralize splitArtistString
